moved the pourer attaching to its own file moved test objects to a testObject folder under tests added debug to requiremocker
