+++ lib/RequireMocker.js
@@ -1,4 +1,5 @@
 var Module = require("module"),
+    debug = require("debug")("mocha-stirrer:RequireMocker"),
     utils = require("./utils"),
     resolvePath = Module._resolveFilename,
     loadModule = Module._load;
@@ -41,7 +42,7 @@ var RequireMocker = (function () {
 
         options = options || {};
 
-        //todo: - debug - console.log("resolving path for: " + requirePath + " - parent: " + parent.id);
+        debug("resolving path for: " + requirePath + " - parent: " + parent.id);
 
         var orgRequire = Module.prototype.require,
             requirePathResolved = resolvePath(requirePath, parent); //resolve the path as if we required from the parent
@@ -89,7 +90,7 @@ var RequireMocker = (function () {
             var resolvedPath = Module._resolveFilename(path, this);
 
             if (path === modulePath || this.id !== modulePath) {    //dont mock if this is the test module or if this require is not from the test module
-                //todo: - debug - console.log(" $$$$$$$$$$$$$$$$$ ---- REAL !!!!!! " + path);
+                debug("requiring module as is - not stubbing: " + path);
 
                 if (path === modulePath) {
                     _removeFromNodeCache(resolvedPath); //need to remove the tested module from node's cache every time its mock-required
@@ -98,8 +99,6 @@ var RequireMocker = (function () {
                 reqResult = orgRequire.call(this, path);  //do the real require for the module being tested
             }
             else {
-
-
                 var stubResult = _stubRequiredModule(mocker, path, resolvedPath, options, this);
 
                 _addToStubs(mocker._stubs, resolvedPath, stubResult);
@@ -144,18 +143,13 @@ var RequireMocker = (function () {
     function _restore(mocker, full) {
 
         if (mocker._modules) {
-
-            //todo: - debug - console.log(" $$$$$$$$$$$$$$$$$ ---- clearing modules !!!!!! full = " + full, mocker._modules);
-
             mocker._modules.forEach(function (path) { //need to remove the modules that were mock-required from node's cache so they can be mock-required again
                 _removeFromNodeCache(path);
             });
         }
 
         if (mocker._stubs) {
-
-            //todo: - debug - console.log(" $$$$$$$$$$$$$$$$$ ---- clearing stubs !!!!!! full = " + full, Object.keys(mocker._stubs));
-
+            debug("restoring stubs");
             Object.keys(mocker._stubs).forEach(_restoreStub.bind(mocker, full));
         }
 
@@ -164,10 +158,11 @@ var RequireMocker = (function () {
 
     function _restoreStub(full, id) {
 
-        var stub = this._stubs[id];
+        if (full) {
+            var stub = this._stubs[id];
 
-        if (stub) {
-            if (full) {
+            if (stub) {
+                debug("restoring stub: " + id);
                 if (stub.prototype) {
                     _restoreProperties(stub.prototype);
                 }
@@ -221,7 +216,8 @@ var RequireMocker = (function () {
             }
 
             if (!_dontStub(path, resolvedPath, options.dontStub)) { //didnt find in the dont stub list - so stubbing!
-                //todo: - debug - console.log(" $$$$$$$$$$$$$$$$$ ---- STUBBING !!!!!!! - " + resolvedPath);
+
+                debug("stubbing mock-required module: " + resolvedPath);
 
                 if (utils.isFunc(moduleExports)) {
                     if (moduleExports.prototype && !utils.isObjectEmpty(moduleExports.prototype)) {
@@ -272,7 +268,6 @@ var RequireMocker = (function () {
     function _addToStubs(stubs, resolvedPath, stubResult) {
 
         if (stubResult.stub) {
-            //todo: - debug - console.log("$$$$$$$$$$$$$$$$$ ---- adding to stubs with id: " + resolvedPath);
             stubs[resolvedPath] = stubResult.stub;
         }
     }
@@ -284,6 +279,7 @@ var RequireMocker = (function () {
             var setupFn = options.setup[path] || options.setup[resolvedPath];
 
             if (setupFn) {
+                debug("running setup function for: " + resolvedPath);
                 setupFn(stub);
             }
         }
@@ -300,5 +296,5 @@ var RequireMocker = (function () {
     return RequireMocker;
 })();
 
-
+//******************** EXPORT ********************
 module.exports = RequireMocker;
