+++ lib/stirrer.js
@@ -1,10 +1,14 @@
 var sinon = require("sinon"),
     debug = require("debug")("mocha-stirrer:stirrer"),
+    pourer = require("./pourer"),
     Mocker = require("./RequireMocker"),
     utils = require("./utils");
 
-//todo: define default for setup/teardown on before/after or beforeEach/AfterEach
 //todo: implement clear method to remove befores/afters/pars/fakes
+//todo: consider adding 'full' parameter to restir method to clear pars, befores, afters
+//todo: implement delay grind par to do setup manually
+//todo: inject the stubs from mocker to the cup stubs
+//todo: implement requires which can be part of the conf of the stir method
 
 var stirrer = (function () {
     "use strict";
@@ -16,12 +20,6 @@ var stirrer = (function () {
     });
 
     var EMPTY = "-_empty_-";
-    var TEST_TYPES = {
-        NORMAL: 1,
-        ONLY: 2,
-        SKIP: 3,
-        WRAP: 4
-    };
 
     /**
      *
@@ -31,16 +29,17 @@ var stirrer = (function () {
      *      - stubs (optional) - object or a function
      *      - spies (optional)
      *      - mocks (optional)
+     *      - delay (default: false) - Boolean
      *      - transform (optional) - function
-     *      - before (optional)
-     *      - after (optional)
-     *      - beforeEach (optional)
-     *      - afterEach (optional)
-     *      - sandbox (optional) - configuration for sinon sandbox
-     *      - setupImmediate (default: false)
-     *      - transformForEach (default: false)
-     *      - dontRestir (default: false)
-     *      - requires (optional)
+     *      - before (optional) - function
+     *      - after (optional) - function
+     *      - beforeEach (optional) - function
+     *      - afterEach (optional) - function
+     *      - sandbox (optional) - configuration object for sinon sandbox
+     *      - setupImmediate (default: false) - Boolean
+     *      - transformForEach (default: false) - Boolean
+     *      - dontRestir (default: false) - Boolean
+     *      - requires (optional) - object or a function
      *      - befores (optional) - function or array of functions with signature: fn(next)
      *      - afters (optional) - function or array of functions with signature: fn(next)
      * @param testFn
@@ -67,15 +66,13 @@ var stirrer = (function () {
         return cup;
     }
 
-    function restir(cup) { //todo: consider adding 'full' parameter to clear pars, befores, afters
+    function restir(cup) {
         cup.sb.restore();
     }
 
-    function mockRequire(cup, requirePath, options) {
+    function fakeRequire(cup, requirePath, options) {
 
         var reqModule = cup._mocker.require(module.parent, requirePath, options);
-
-        //todo: inject the stubs from mocker to the cup stubs
 
         _enrichCupStubsFromMocker(cup);
 
@@ -150,7 +147,7 @@ var stirrer = (function () {
         return {
             require: function (path, options) {
                 debug("cup require called with path = " + path + " - " + this.name);
-                return mockRequire(this, path, options);
+                return fakeRequire(this, path, options);
             },
 
             stir: function (conf) {
@@ -165,82 +162,15 @@ var stirrer = (function () {
         };
     })();
 
-    function _attachPour(cup, conf) {    //todo: move pour to an external file
-
-        function internalPour(testName, testFn, type) {
-
-            testName = testName || "stirrer test - " + cup.name;
-            type = type || TEST_TYPES.NORMAL;
-
-            //_runAfters(cup, conf);
-
-            function pourOwnTestFn(done) {
-
-                var isAsync = type !== TEST_TYPES.SKIP &&
-                    type !== TEST_TYPES.WRAP &&
-                    utils.isAsync(testFn, 0);
-
-                _runBefores(cup, function () {
-
-                    var ownDone = function (err) {//need to wrap the done before giving it to the test fn
-                        _runAfters(cup, function () {   //run any afters before signaling we're done with the test
-                            done(err);
-                        });
-                    };
-
-                    if (isAsync) {
-                        testFn.call(cup, ownDone); //make the cup as the context
-                    }
-                    else {
-                        testFn.call(cup);  //make the cup as the context
-                        ownDone(); //done(); //need to call done because the passed test fn didnt mean to call it (no arg declared)
-                    }
-                });
-            }
-
-            if (utils.isFunc(testFn)) {
-                switch (type) {
-                    case TEST_TYPES.ONLY:
-                        it.only(testName, pourOwnTestFn);
-                        break;
-                    case TEST_TYPES.SKIP:
-                        it.skip(testName, pourOwnTestFn);
-                        break;
-                    case TEST_TYPES.NORMAL:
-                        it(testName, pourOwnTestFn);
-                        break;
-                    case TEST_TYPES.WRAP:
-                        pourOwnTestFn(_noop);//pour is used just as a wrapper, the Mocha 'it' should be called by the testing code
-                        break;
-                }
-            }
-            else {
-                it(testName); //define a pending test
-            }
-        }
-
-        internalPour.only = function (testName, testFn) {
-            internalPour(testName, testFn, TEST_TYPES.ONLY);
-        };
-
-        internalPour.skip = function (testName, testFn) {
-            internalPour(testName, testFn, TEST_TYPES.SKIP);
-        };
-
-        internalPour.wrap = function (testName, testFn) {
-            internalPour(testName, testFn, TEST_TYPES.WRAP);
-        };
-
-        cup.pour = internalPour;
-    }
-
     function _blend(conf) {
 
         var name = conf.name || Date.now();
         var sbConf = conf.sandbox || void 0;
         var sb = sinon.sandbox.create(sbConf);
 
-        _setHooks(conf);
+        if (!conf.delay) {
+            _setHooks(conf);
+        }
 
         var cup = {
             name: name,
@@ -251,10 +181,8 @@ var stirrer = (function () {
             _afters: []
         };
 
-        //todo: implement requires which can be part of the conf
-
         utils.merge(cupBody, cup);
-        _attachPour(cup, conf);
+        pourer.attach(cup);
         cup.stir(conf);
 
         return cup;
@@ -355,17 +283,6 @@ var stirrer = (function () {
                 finishAfter();
             }
         });
-
-        //if (utils.isAsync(conf.after)) {
-        //    after(name, function (done) {
-        //        conf.after(cup, done);
-        //    });
-        //}
-        //else {
-        //    after(name, function () {
-        //        conf.after(cup);
-        //    });
-        //}
     }
 
     function _doAfterEach(cup, conf) {
@@ -382,32 +299,6 @@ var stirrer = (function () {
                 conf.afterEach(cup);
             });
         }
-    }
-
-    function _runBefores(cup, callback) {
-
-        debug("running cup befores functions count = " + cup._befores.length + " - " + cup.name);
-
-        utils.runSeriesAsync(cup._befores, function () {
-            callback();
-        }, cup);
-    }
-
-    function _runAfters(cup, callback) {
-
-        //after("stirrer test afters - " + cup.name, function (done) {
-
-        debug("running cup afters functions count = " + cup._afters.length + " - " + cup.name);
-
-        utils.runSeriesAsync(cup._afters, function () {
-
-            //if (!conf.dontRestir) {
-            //    cup.restir(); //make sure all fakes are restored using sinon's sandbox restore
-            //}
-
-            callback();//done();
-        }, cup);
-        //});
     }
 
     function _setup(cup, conf) {
@@ -527,7 +418,7 @@ var stirrer = (function () {
         create: grind, //alias (for those that dont get it)
         restir: restir,
         reset: restir, //alias (for those that dont get it)
-        require: mockRequire,
+        require: fakeRequire,
 
         // Making the RequireMocker class available to the outside so it can be used as standalone
         RequireMocker: Mocker,
