+++ test/RequireMocker.standalone.test.js
@@ -16,7 +16,7 @@ describe("testing auto mocking for require with standalone mocker", function () 
 
     it("requiring the module should normally should work normally", function(){
 
-        var foo = require("./foo");
+        var foo = require("./testObjects/foo");
 
         expect(foo).to.exist();
         expect(foo.bar()).to.equal("foo");
@@ -30,7 +30,7 @@ describe("testing auto mocking for require with standalone mocker", function () 
 
         it("should mock require successfully - standalone", function () {
 
-            var foo = mocker.require("./foo", {
+            var foo = mocker.require("./testObjects/foo", {
                 dontStub: ["fs"]
             });
 
@@ -41,7 +41,7 @@ describe("testing auto mocking for require with standalone mocker", function () 
             expect(foo.useSubDep("world")).to.not.exist();
             expect(foo.useFuncDep()).to.not.exist();
 
-            var Bar = require("./sub/bar");
+            var Bar = require("./testObjects/sub/bar");
 
             expect(Bar.prototype.useDep).to.have.been.calledWith("world");
 
@@ -53,7 +53,7 @@ describe("testing auto mocking for require with standalone mocker", function () 
 
         it("should mock require and setup stub successfully - standalone", function () {
 
-            var foo = mocker.require("./foo", {
+            var foo = mocker.require("./testObjects/foo", {
                 setup: {
                     "./sub/bar": function (stub) {
                         //set up the stub function to return a static string
@@ -72,7 +72,7 @@ describe("testing auto mocking for require with standalone mocker", function () 
             expect(foo.useSubDep("world")).to.equal("this works!");
             expect(foo.useFuncDep()).to.equal("bla bla");
 
-            var Bar = require("./sub/bar");
+            var Bar = require("./testObjects/sub/bar");
 
             expect(Bar.prototype.useDep).to.have.been.calledWith("world");
 
@@ -82,7 +82,7 @@ describe("testing auto mocking for require with standalone mocker", function () 
 
     it("requiring the same module normally again should now work normally still", function(){
 
-        var foo = require("./foo");
+        var foo = require("./testObjects/foo");
 
         expect(foo).to.exist();
         expect(foo.bar()).to.equal("foo");
@@ -94,7 +94,7 @@ describe("testing auto mocking for require with standalone mocker", function () 
 
     it("requiring a module that was a stub in previous should now work normally - not stubbed - standalone", function () {
 
-        var Bar = require("./sub/bar");
+        var Bar = require("./testObjects/sub/bar");
         var bar = new Bar();
         expect(bar.start()).to.equal("hello world");
         expect(bar.getStats()).to.exist();
