+++ test/RequireMocker.test.js
@@ -12,7 +12,7 @@ describe("testing auto mocking for require", function () {
 
     it("requiring the module should normally should work normally", function () {
 
-        var foo = require("./foo");
+        var foo = require("./testObjects/foo");
 
         expect(foo).to.exist();
         expect(foo.bar()).to.equal("foo");
@@ -30,7 +30,7 @@ describe("testing auto mocking for require", function () {
 
         cup.pour("should mock require successfully", function (done) {
 
-            var foo = cup.require("./foo", {
+            var foo = cup.require("./testObjects/foo", {
                 dontStub: ["fs"]
             });
 
@@ -41,7 +41,7 @@ describe("testing auto mocking for require", function () {
             expect(foo.useSubDep("world")).to.not.exist();
             expect(foo.useFuncDep()).to.not.exist();
 
-            var Bar = require("./sub/bar");
+            var Bar = require("./testObjects/sub/bar");
 
             expect(Bar.prototype.useDep).to.have.been.calledWith("world");
 
@@ -59,7 +59,7 @@ describe("testing auto mocking for require", function () {
 
         cup.pour("should mock require and setup stub successfully", function () {
 
-            var foo = cup.require("./foo", {
+            var foo = cup.require("./testObjects/foo", {
                 setup: {
                     "./sub/bar": function (stub) {
                         //set up the stub function to return a static string
@@ -78,7 +78,7 @@ describe("testing auto mocking for require", function () {
             expect(foo.useSubDep("world")).to.equal("this works!");
             expect(foo.useFuncDep()).to.equal("bla bla");
 
-            var Bar = require("./sub/bar");
+            var Bar = require("./testObjects/sub/bar");
 
             expect(Bar.prototype.useDep).to.have.been.calledWith("world");
         });
@@ -86,7 +86,7 @@ describe("testing auto mocking for require", function () {
 
     it("requiring the same module normally again should now work normally still", function () {
 
-        var foo = require("./foo");
+        var foo = require("./testObjects/foo");
 
         expect(foo).to.exist();
         expect(foo.bar()).to.equal("foo");
@@ -98,7 +98,7 @@ describe("testing auto mocking for require", function () {
 
     it("requiring a module that was a stub in previous should now work normally - not stubbed", function () {
 
-        var Bar = require("./sub/bar");
+        var Bar = require("./testObjects/sub/bar");
         var bar = new Bar();
         expect(bar.start()).to.equal("hello world");
         expect(bar.getStats()).to.exist();
@@ -118,5 +118,9 @@ describe("testing auto mocking for require", function () {
         var resolved = path.resolve("/");
         expect(resolved).to.equal("/");
     });
-});
 
+    describe("use stirrer cup with requires (mocker)", function () {
+
+        it("need to test");
+    });
+});
