+++ test/testObjects/foo.js
