+++ test/testObjects/sub/bar.js
