+++ test/testObjects/sub/func.js
